+++ src/client/src/Domains/FacultyMember/api/api.ts
@@ -1,49 +1,7 @@
 import API from 'api';
 
-export const departments = [
-    {
-        label: 'Biochemistry',
-        value: 'Biochemistry',
-    },
-    {
-        label: 'Bioengineering',
-        value: 'Bioengineering',
-    },
-    {
-        label: 'Chemical Engineering',
-        value: 'Chemical Engineering',
-    },
-    {
-        label: 'Computer Science',
-        value: 'Computer Science',
-    },
-    {
-        label: 'Computer Engineering',
-        value: 'Computer Engineering',
-    },
-    {
-        label: 'Data Science',
-        value: 'Data Science',
-    },
-    {
-        label: 'Environmental Engineering',
-        value: 'Environmental Engineering',
-    },
-    {
-        label: 'Electrical Engineering',
-        value: 'Electrical Engineering',
-    },
-    {
-        label: 'Materials Science and Engineering',
-        value: 'Materials Science and Engineering',
-    },
-    {
-        label: 'Mechanical Engineering',
-        value: 'Mechanical Engineering',
-    },
-];
 
-export interface IFacultyMemberProfile {
+export interface IFacultyMemberProfileForm {
     firstName: string; 
     middleName: string; 
     lastName: string; 
@@ -56,7 +14,7 @@ export interface IFacultyMemberProfile {
 }
 
 //TODO: DEFINE THIS LATER!
-export async function createFacultyMemberProfile(facultyMemberProfile: IFacultyMemberProfile) {
+export async function createFacultyMemberProfile(facultyMemberProfile: IFacultyMemberProfileForm) {
     return API.post('facultyMemberProfile/create', {
         facultyMemberProfile: {
             ...facultyMemberProfile,
