+++ src/client/src/sharedData.ts
@@ -1,0 +1,42 @@
+export const departments = [
+    {
+        label: 'Biochemistry',
+        value: 'Biochemistry',
+    },
+    {
+        label: 'Bioengineering',
+        value: 'Bioengineering',
+    },
+    {
+        label: 'Chemical Engineering',
+        value: 'Chemical Engineering',
+    },
+    {
+        label: 'Computer Science',
+        value: 'Computer Science',
+    },
+    {
+        label: 'Computer Engineering',
+        value: 'Computer Engineering',
+    },
+    {
+        label: 'Data Science',
+        value: 'Data Science',
+    },
+    {
+        label: 'Environmental Engineering',
+        value: 'Environmental Engineering',
+    },
+    {
+        label: 'Electrical Engineering',
+        value: 'Electrical Engineering',
+    },
+    {
+        label: 'Materials Science and Engineering',
+        value: 'Materials Science and Engineering',
+    },
+    {
+        label: 'Mechanical Engineering',
+        value: 'Mechanical Engineering',
+    },
+];
